feat: Update pubspec.yaml and AndroidManifest.xml for launcher icon
This commit updates the `pubspec.yaml` file to replace the `flutter_icons` dependency with `flutter_launcher_icons` and configures it to use the `launcher_icon` asset for the Android platform. Additionally, the `AndroidManifest.xml` file is updated to reference the new launcher icon. This change ensures that the app uses the desired launcher icon on Android devices.

Recent user commits:
- change icon

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression