feat: Update dependencies and remove unused code
This commit updates the dependencies for the Linux and Windows platforms, ensuring compatibility and functionality improvements. Additionally, it removes the `image_gallery_saver` dependency, as it was no longer being used in the codebase. These changes optimize the project by reducing unnecessary dependencies and keeping the codebase up to date.

Recent user commits:
- change icon
- feat: Remove image_gallery_saver dependency
- feat: Update dependencies for Linux and Windows platforms
- feat: Add check for building name in add_lost_thing screen
- feat: Delete lost thing image from Firebase Storage if it exists

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression