feat: Add image URL to post data
This commit adds the image URL to the post data in the `PostProvider` class. The `imageUrl` property of the `updatedPost` object is now included in the data sent to the server. This change ensures that the image associated with the post is properly stored and displayed.

Recent user commits:
- change icon

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression