feat: Update edit post screen to display selected image or URL
This commit updates the `_EditPostPageState` class in the `edit_post_screen.dart` file. It modifies the code to display either the selected image or the image URL, depending on the availability of the data. This change improves the user experience by ensuring that the image associated with the post is properly displayed.

Recent user commits:
- change icon

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression