feat: Improve image display in edit post screen
This commit updates the `_EditPostPageState` class in the `edit_post_screen.dart` file. It improves the logic for displaying images by checking if the image path and image bytes are not empty, instead of checking if they are empty. This change ensures that the selected image or URL is properly displayed in the edit post screen.

Recent user commits:
- change icon

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression