feat: Improve delete account functionality
This commit simplifies the logic and removes unnecessary code in the `_DeleteAccountScreenState` class of the `delete_account_screen.dart` file. It also adds the functionality to log out the user and clear the stored account and password information from shared preferences. After successfully deleting the account, a success message is displayed using the `showAlertDialog` function. The user is then redirected to the login screen.

Recent user commits:
- feat: Improve image display logic in edit post screen

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression