feat: Update surface color in light theme and search bar background color in main page
This commit updates the surface color in the light theme of the `main.dart` file to use a lighter shade of grey (Colors.grey[300]). It also updates the background color of the search bar in the `main_page.dart` file to match the surface color of the current theme. These changes improve the visual consistency and user experience in the app.

Recent user commits:
- feat: Update image upload logic in edit post screen
- feat: Update chat message widget to display author's message in different color
- feat: Update post type labels to use Chinese language

Recent repository commits:
- feat: Remove web-specific code for setting external user ID
- feat: Enable OneSignal login for non-web platforms
- feat: Add flutter_image_compress dependency for image compression